+++ crates/cli/src/wasm_bindgen_test_runner.rs
@@ -208,6 +208,18 @@ fn rmain(cli: Cli) -> anyhow::Result<()> {
 
     let tmpdir = tempfile::tempdir()?;
 
+    // Support a WASM_BINDGEN_KEEP_TEST_BUILD=1 env var for debugging test files
+    let tmpdir_path = if env::var("WASM_BINDGEN_KEEP_TEST_BUILD").is_ok() {
+        let path = tmpdir.keep();
+        println!(
+            "Retaining temporary build output folder: {}",
+            path.to_string_lossy()
+        );
+        path
+    } else {
+        tmpdir.path().to_path_buf()
+    };
+
     let module = "wasm-bindgen-test";
 
     // Right now there's a bug where if no tests are present then the
@@ -342,15 +354,15 @@ fn rmain(cli: Cli) -> anyhow::Result<()> {
     b.debug(debug)
         .input_module(module, wasm)
         .emit_start(false)
-        .generate(&tmpdir)
+        .generate(&tmpdir_path)
         .context("executing `wasm-bindgen` over the Wasm file")?;
     shell.clear();
 
     match test_mode {
         TestMode::Node { no_modules } => {
-            node::execute(module, tmpdir.path(), cli, tests, !no_modules, coverage)?
+            node::execute(module, &tmpdir_path, cli, tests, !no_modules, coverage)?
         }
-        TestMode::Deno => deno::execute(module, tmpdir.path(), cli, tests)?,
+        TestMode::Deno => deno::execute(module, &tmpdir_path, cli, tests)?,
         TestMode::Browser { .. }
         | TestMode::DedicatedWorker { .. }
         | TestMode::SharedWorker { .. }
@@ -365,7 +377,7 @@ fn rmain(cli: Cli) -> anyhow::Result<()> {
                 },
                 headless,
                 module,
-                tmpdir.path(),
+                &tmpdir_path,
                 cli,
                 tests,
                 test_mode,